requirements.txt
Django
django-jsonsuit
# -e ../django-jsonsuit
requests